Updated numbering for TOC and body titles
Changed to enable the numbering of embedded topics. Previously, topics which contained themselves multiple other topics were left unnumbered in the final output because they had no presence in the opentopics:map element of the consolidated xml. I've changed the commons.xsl file to use the main body of the .xml for calculating number values and just to use the opentopics:map element as a reference point to avoid numbering frontmatter content and to assign letters instead of numbers to topic-level appendix topics.